// src/components/infoBoxTotals/interfaces.ts
export interface IStationTotals {
  bikes: number;
  electrical: number;
  mechanical: number;
  docks: number;
}

export interface INumbersHighlighting {
  isBikesNumberHighlighted: boolean;
  isElectricalBikesNumberHighlighted: boolean;
  isMechanicalBikesNumberHighlighted: boolean;
  isDocksNumberHighlighted: boolean;
}